hysteresis feature added:
ip will be added in to the "tcp forced" list when the counter of requests per period is equal or more than threshold and ip will be deleted when the counter is less than threshold - hysteresis. Default hysteresis is threshold /10
It is possible to disable releasing - just set hysteresis = threshold. In this case ip will be added to "tcp forced" list forever